Retry mounting the shares up to 10 times.
There's a race between the system mounting the share and the findmnt invocation in the script to determine if the share is already mounted.
When findmnt wins the race the script would previously operate as if the share definitely needed to be mounted, and would fail to mount it.
Now the script will notice that the share has already been mounted.